Add some SVG version for the current default set
These are SVGs from the Tango icon set as well as some theme-independent self-made icons. Many icons are missing. This set s a wild mixture of Tango, old KDE icons and some self-made icons drawing on (to me) unknown models.

I leave the task to finish this set (and thus the SVG icons generally) to others.