Fix overflow of inset over text (see #9224)
The symptoms are the same as the original #9224, but the reason is different. In the new code, the width of justified rows includes the Row::separator value. Therefore the CRC computed in ParagrahMetrics::computeRowSignature is not updated in some cases.

The fix is to add Row::separator as one of the elements of the row crc.