installer: configure.nsh: better babel installation
The former solution with installing all babel packages at once via a list fails if only a single package of the list is already installed. In this case none of the packages in the installation list will be installed. Therefore every package will be installed separately.

(The bug is already reported to MiKTeX and this workaround can be removed for LyX 2.1.4.))