Add a "Development policies" section to dev manual
Writing these policies down makes it easier for newcommers to
understand guidelines and the intuition behind them.

The two policies this commit adds were responses by Georg and Richard
to two threads on the mailing list. The policies are titled
"When to set a fixed milestone" and
"Can we add rc entries in stable branch?".