Move svg icons to proper place.
You have to manually delete the lib/images/svg directory as git
tracks only files.

There may be warnings on the terminal due to some faulty icons that have
to be found and corrected. Moreover, some of the icons are not rendered
too well at the default size. These ones should be properly tweaked.
Kudos to Jürgen for putting together such a large number of icons.